BUG#35507456 -  Assertion `[&]() { for (const Predicate *first = begin; first < end; first++)
The above bug is a regression introduced by
"Change-Id: I6cb8af97c30870d36bd8b0fb283dbd954ca735e1"
(fix for Bug#34763224).

That commit ordered predicates using a comparison function that was
not truly transitive (so that in some cases all of the following could
be true: p1<p2, p2<p3, p3<p1).  This would trigger an assert that
verified the ordering by comparing all pairs of predicates.

This commit fixes that by using a comparison function that is
transitive.

Change-Id: I835be0fb9f1de59354c2ce0e25b91a9570dd03a9