Bug#35522631 wrong error-log at test-failure in integration tests
When integration tests start need multiple servers they prepare
a datadir once and copy the files to the new servers to speed up
starting server instance as the initial setup is costly.

Currently, the errorlog is copied which leads to other instances
wrongly showing the errorlog of the initialize step.

Change
------

- delete error-log after prepare-datadir

Change-Id: I890ae2211b0e4229b101ce7212ccc48582667c19