Bug#35522821 setup_mysqld_accounts() breaks tests with replicated environments
setup_mysqld_accounts() is currently called as part of the
"prepare-instance-once, copy often" step of shared-server.

setup_mysqld_accounts() creates users, schemas, tables, functions, ...
needed for all tests. Those statements are written to the binlog, but
currently not replicated.

As soon as replication would be enabled, the same statements would be
execute on the other nodes and lead to a replication break as the same
functions, tables, ... already exist.

Change
------

- don't call setup_mysqld_accounts() in the prepare-server step
- call from the test-suite if they know they don't want to replicate

Change-Id: I23e329773c47ef8f17891147e6f7791af82b7118