Bug#34388068 NDB: Update to char data type columns in primay keys are lost on replica
Post push patch, fixing the MTR test ndb_binlog_variants.test

Order of binlog  rows selected from binlog_stmts table were not
deterministic due to case insensitive collation were used
for the 'txt' column in that table.

Thus, when ordering selected rows on 'txt', it was not deterministic
in which order the upper or lower case valued rows appeared in.

Patch define the txt column with the case sensitive utf8mb24_bin
collation. It also reorder the columns in the cpk table such that
the numeric 'nonchar' column appeared first as the binlog result is
expected to be ordered on that column.

Change-Id: If8cb06e106cd3b9ac6cc70696679cf4c9df06795