Bug#34900735 - Integer overflow while printing query_id
Problem:
Query ID (uint32_t) was being incorrectly logged with a %d format
specifier instead of the correct %u.

Example:

 "ts":1671103676877,
 "component": "QCOMP",
 "filename": "QCOMP_qep_dump_3221225488_01.json", <-- correct format
 "query_id": -1073741808,                         <-- incorrect format
  ...

Solution:
Add -Wformat-signedness in handler and rpdmaster code when compiling
with GCC in order to systematically fix the formatting issues.

This patch also fixes many formatting issues in olrapid/ code but does
not add the flags there as hundreds of issues still remain.

Change-Id: I45a32ffae61f8a52cddeee18af9b022781748497