WL#15154 patch #5 Establish TLS
On the client side, in Transporter::connect_client(), check the
TLS negotiation result, initiate the TLS handshake, and perform
certificate authorization checks.

On the server side, in TransporterService::newSession(), check the
TLS negotiation result and complete the TLS handshake. Then in
TransporterRegistry::connect_server(), after receiving the client's
hello message, perform certificate authorization checks.

The connect and disconnect impls in TCP_Transporter are responsible
for maintaining the m_encrypted flag and the certificate table. In
TCP_Transporter::connect_common(), locking is enabled around the
SSL object; the lock must be acquired in order to send data, receive
data, or close the socket.

Change-Id: I9758adc88e8f152760bab3df1485556681e88d07