Bug#35493481: Forced secondary engine execution goes to primary engine and picks wrong error.
The problem occurs when execution is forced to go to a secondary
storage engine. After such execution, MySQL executor sets up for
execution in the primary engine by setting secondary_engine_optimization
as PRIMARY_ONLY. However, when execution is attempted in the primary
engine, the executor sees that execution is forced secondary.
It will then attempt to look up the secondary engine, retrieve the
last status from the secondary engine and return this as the result
of statement execution.

The problem with this strategy is that sometimes an incorrect error code
is returned. It seems more logical to return directly after execution
in secondary engine is performed, without going to the primary engine.

The following changes are implemented in this patch:

- When FindBestQueryPlanInner() is asking a secondary engine to optimize
  a statement and no optimized plan is returned, the function will pick
  up the current engine fail reason and return that along with error
  code ER_SECONDARY_ENGINE (3889) as the outcome of the optimization.
  This error code is later picked up by handlers for regular and
  prepared statements when statement execution is forced secondary.

  A similar adjustment was made to the function NotSupportedInRapid().

- In check_secondary_engine_statement(), if a statement was forced to
  secondary engine, just exit with the error.
  check_secondary_engine_statement() is also called only if there is
  an error in the execution.

- In Prepared_statement::execute_loop(), if a statement that was
  executed in a secondary engine failed, only re-execute it in the
  primary engine if secondary engine execution is not forced.

- The logic in validate_use_secondary_engine() that decides if a
  statement should be executed in the secondary engine is moved to
  open_secondary_engine_tables(), thus it is performed as early as
  possible.

It was also needed to change Sql_cmd_dml::prepare() so that secondary
engine_execution_context is cleared also after a statement from stored
procedure is prepared, in addition to prepared statements.

Additional fix: mysqld_stmt_prepare() no longer saves and restores
the secondary engine optimization, since the value is never changed
inside that function.

Additional fix: No longer check thd->is_fatal_error() after parsing
because it may not be synchronized with diagnostics area, and thus
we might try to peek at a diagnostics area with a "success" status.

Test changes:

- Several tests now report error issued by MySQL optimizer instead of
  ER_SECONDARY_ENGINE (3889). This seems like a real improvement since
  the statement is rejected long before attempting to offload it to
  HeatWave. An example is cp_i_subquery_dict and cp_i_subquery_varlen
  that now report ER_SUBQUERY_TRANSFORM_REJECTED (3973).

- if ($PS_PROTOCOL) is no longer needed in many tests that are run
  with regular protocol and --ps-protocol. This is possible because
  code path for regular and prepared statements are now similar.
  An example is found in tests cp_partition_locking_dict and
  cp_partition_locking_varlen.

- Several test result error messages are now more specific. One example
  is from analytics_query_dict where the old message was
  "use_secondary_engine is FORCED but query could not be executed in
  secondary engine" whereas the new message is
  "Unsupported argument type for LIKE".

- By exiting directly after a failed forced secondary engine execution,
  optimizer trace handling can be simplified. It is no longer necessary
  to set optimizer_trace_offset to -2, it is sufficient to look at the
  most recently traced statement only. For example, see tests
  optimizer_trace_dict and optimizer_trace_varlen.

- Some tests have been adjusted to return ER_SECONDARY_ENGINE_PLUGIN
  (3877) instead of ER_SECONDARY_ENGINE (3899), however we have tried
  to minimize such changes as much as possible, while still having
  minimal code changes.

Change-Id: Ifd54b6ecc19d795db73e9d763bba2177004ba09f