Bug#35503286 MySQL Router: routing through named socket(s) won't resume after cluster recovery
Each routing section can be configured to accept the client connections:
- using tcp port
- using unix socket (--conf-use-sockets --conf-skip-tcp)
- using both (--conf-use-sockets)

WL#13327 "accept connections only if destinations available" introduced
a functionality when the accepting ports and sockets are closed when there
are not destination available. They also should be reopened each time the
destination(s) are available again. This was not the case for the unix socket
acceptor.

This patch:
- makes sure the unix socket acceptors are also reopened once there are
  available routing destinations
- removes the unix socket file when there are no available destinations,
  similar way the accepting port is closed for TCP acceptors
- recreates the unix socket file once there are destinations available

Change-Id: I5913e0768f4c20c19314bc77a213872943c0b805