Bug#35441583 Cache spill checker prints two warnings in a row
Problem:

On commit of the binlog transaction, the binlog thread checks if some
binlog contents had to be written to a temporary file because the
cache buffers was full; called a disk spill. If there are disk spills,
the thread prints a warning to the mysqld log suggesting an increase
of the binlog cache size. The printing of this warning is
rate-limited, because a transaction can commit every 100ms and a
single warning with this frequency is undesirable. The rate-limiting
code, however, allows two consecutive warnings.

Analysis:

The code is using a mask to bound warnings, by comparing the current
number of disk spills (m_disk_spills) with the mask. Additionally, if
the m_disk_spills is higher than the mask's value, it performed a
modulus operation to bound the warnings. Considering that every binlog
transaction has disk spills, this creates an issue that if the mask is
a high value, then the warnings will only be printed every (mask *
0.1) seconds. The current value results in almost half-hour before a
warning comes and, hence, "mask" number of disk spilled transactions
before the user is warned.

Solution:

Modify existing code to have more frequent warnings and avoid
duplicate warnings; e.g. every 64th time, translating into ~= 6.4
seconds on stable throughput.

Change-Id: I2f31e10b8d74b5453204dc0309bdc8bf1cce25e8