Bug#35550389 Add imported libraries for system protobuf
Clean up the cmake code for system/bundled protobuf.

The most recent versions of protobuf depend on google abseil, so we
will need a clean way of specifying library and interface
dependencies.  See Bug #111469 Build failure with Protobuf 22+

This patch also removes some code using google::protobuf::LogHandler,
which is not available in protobuf 22+
The code was for internal debugging only, and rarely used.

Change-Id: I1567f0a6a25f24ba1384f8c192a11c99339f9513