Bug#35338567 Role with backslash in grant does not allow proper SHOW DATABASES
Problem:
When a backslash is used to escape the wildcard '_' in database
name pattern inside GRANT statement, the SHOW DATABASES and USE
<Database_name> queries do not work correctly for the user.

Fix:
For SHOW DATABASES:
CAN_ACCCESS_DATABASE() is fixed to call db_acl() specifying that
the db name is not a pattern string when there are active roles.
For USE <Database_name>:
mysql_change_db() is fixed to specify that db name is not a
pattern string when it calls ab_acl() in case there are active
roles.

Change-Id: I7f14ef4eb9f02c195b10f5506cc511774424975e