Bug#35471453: MySQL debug server stops when executing query
The error generation code was not prepared to handle CREATE/DROP ROLE
as top level statements when checking for orphan SQL objects (views
in this case). Added a case for the two new commands and also
moved the code into the branch that needs it.
Fixed the line endings of a test case.
Added a test.

Change-Id: I5ffbf3fc816529d506fd8e1a26a00917c1df1819