Bug#36083050 Run a single Cluster/J test case
In the Cluster/J test runner, AllTests::main(), accept the command-line
option -n followed by a list of test names. In this case, only run
the named tests.

Also support a new option --print-cases (short form -l) option,
to list tests.

In the Cluster/J test runner (TestRunner.java), fix a spelling
error in the output, and if no tests were run print "no tests run"
rather than "all tests succeeded".

Add a new NDBAPI test utility called testClusterJ. This program takes
care of creating a clusterj.properties file, constructing the Java
classpath and command line, and executing the test runner. It takes no
command-line options of its own; all supplied options are passed
verbatim to the Java program. If MTR_CLASSPATH is set in the
environment, this is appended to the classpath.

The clusterj.properties file is created in TMPDIR, and is deleted
when the program exits.

If NDB_CONNECTSTRING is set in the environment, testClusterJ uses it
in clusterj.properties. If not, and a management server is listening
at port 13000 (as in the common case of an MTR environment), it uses
port 13000. Otherwise it falls back to port 1186.

If CLUSTERJ_MYSQLD is set in the environment, testClusterJ uses it
in the JDBC connect string in clusterj.properties. If not, and
a management server was found at port 13000, it uses "localhost:13001".
Otherwise it falls back to "localhost:3306".

Change-Id: Ie42fd37057e8f8afb27639a9216deadf4ab6a2b1