Bug#35508108 Query succeeds with subquery to derived=on when it should throw error
The transformation of correlated scalar subqueries to join with a
derived table fails give expected error if sql_mode has
ONLY_FULL_GROUP_BY set. Transform is only attemped for Heatwave or if
the optimizer_switch has subquery_to_derived set to ON.

We wrap the selected item in an ANY_VALUE if it is not added to the
group by list: since this is a scalar subquery and the query is not
grouped, there will be only one row (lest we error out) in the result
set, the selected item will be functionally dependent on the field(s)
we have in the equality predicate(s). We failed to check that the
query is not explicitly grouped.

Solution: do not wrap the selected item in ANY_VALUE if we have a
grouped subquery before the transform.

This patch also addresses the issues raised by Chaithra in the review
(test case added). We did not do cardinality checking correctly in the
presence of explicit GROUP BY. This patch fixes the issue.  The
cardinality checking is changed to use windowed COUNT(*) for

   a) the original group list, and
   b) each inner field added to the grouping

The (earlier) grouped COUNT check is only used for subqueries that are
not explicitly grouped.

The test suite was also updated: in some cases it was wrong (as
compared to what it would be if we disabled transforms altogether,
wrt. cardinality error/no error).

Change-Id: I6c2e9339fdbabb53b837f9287cf3f02f4c814698