Bug#35145188: innodb.log_file_resize_1 is failing sporadically in pb2
innodb_redo_log_capacity_update changes log capacity, and waits for
the next governor thread iteration to finish, but the governor thread
might not actually be aware of the change, as it might have been running in paralell
already. The routine log_files_wait_until_next_governor_iteration
in this case just waits for the current iteration to finish.

This patch alters log_files_wait_until_next_governor_iteration,
to make sure it waits for a full iteration to complete.

Change-Id: Ia359ea371405df3eab1ec5234a86f29d50552250