Bug#35473657 Scalar correlated subquery with group by clause mistakenly transformed [follow-up]
Missing canon update for two RAPID tests:

  group_by_fd_dict
  group_by_fd_varlen

fixed.

Redundant call to Window::resolve_reference causes assert in debug
build of rapid; removed. The reason for the assert was that the
redundant call added the same function twice to Window::m_functions,
and RAPID has an assert that there were no duplicates in the list.

Change-Id: Ic18e7f91dc53bc72ed67e44d6d086ca92ea811ee