Bug#35574407 PB2 failures: Error: Could not register this Router instance
Router component tests often fail on pb2 when the bootstrap is done due to
the DNS issues in this env. The failure is:

Could not get local host address: Name or service not known(errno: -2)
You may want to try --report-host option to manually supply this hostname.

This patch makes sure all the tests that do bootstrap are using --report-host
to avoid DNS query procedure.

Change-Id: I886db73326ee67f6d16aeb75837c863264a270b8