Bug#35574557 connection not disconnected in between recv and write
If mysqlrouter is terminated with SIGTERM/SIGINT it

- disconnects all active connections
- waits until all connections are closed
- exits

If router is between receiving and forwarding a message,
the disconnect request does not lead to a connection close.

From the outside, the router doesn't shutdown until the
client or server connections is closed by the user.

The reason for the hang is that the .disconnect() cancels async
operation, but in case of .async_send() there may be no async operation
to cancel as the write() is synchronous followed by a net::defer()
which isn't cancelled by the .disconnect().

Change
------

- additional to canceling async operations and closing the connection,
  check if a disconnect was requested before starting a async operation.

Change-Id: I516b6226c54b53a0047e2cb8a4e08c6fd939f003