Bug#35575217 MTR: ndb_mgm_show_table.inc prints mgmd(s) nodes in wrong section
Problem:
ndb_mgm_show_table.inc runs 'ndb_mgm -e show' command and mask the
details that are unwanted for the test.
The result of SHOW command is stored into a temporary table and later
parsed and printed to stdout/file.

During the parsing the nodes are grouped into the 3 'types' NDBD(s),
NDB_MGM(s) and MYSQLD(d) based on its id.
(Nodes 1, 2, 3, 4 are mapped to NDBD; node 5 is mapped to NDB_MGM
and nodes 6,7...,19 are mapped to MYSQLD)
This way to map the nodes works for cluster where the node IDs
are as above but for cluster with different configs nodes can be
mapped to a wrong 'type'

Solution:
Group nodes based on their original 'type' returned by SHOW command
instead node id.

Change-Id: Ic5e817e1418ce56a09fdf5778bb0b743ee1af7ef