Bug#35571268 firewall.firewall_long_hostname test run fails
Problem:
The MTR for firewall_long_hostname was failing to run
everytime executed on any platform.

Analysis:
It was found that the problematic area was when the test
was trying to upgrade the firewall tables by starting the
server with the data directory of MySQL 8.0.16, which is
present in the fw_data_80016.zip file, it's available in
mysql-test/std_data folder.

Separately building MySQL 8.0.16 source code and starting
the server with this data directory present in the std_data
folder reveals that there was no shutdown done when the data
directory was created for this test, as the server log is
filled with
[InnoDB] Database was not shutdown normally!
[InnoDB] Starting crash recovery.
[InnoDB] my_realpath(./mtr/#sql-ib1059-165891815.ibd) failed!
[InnoDB] Doing recovery: scanned up to log sequence number 19587304
[InnoDB] Log background threads are being started...
[InnoDB] Applying a batch of 145 redo log records ...
[InnoDB] 10%
[InnoDB] 100% etc.

Fix:
Once the MySQL Server 8.0.16 started after the crash recovery
on the above mentioned data directory, I perform a shutdown
from the client. Then I created a new zip file from this data
directory, and replaced the original file. Test passes now
without any failure after this.

Change-Id: I6888bb5bd8585392aadcdeeab6bc4c3904c3cb9d