Bug#34940000 Hash join execution may be inefficient if probe input is empty
If the probe input of a hash join is empty, we do not need to read the
build input, because we already know that the result will be empty. We
should therefore try to read from the probe input first if:
- The join is a left join or an antijoin. For these join types we must
  read the probe input even if the build input is empty, but the
  converse is not true.
- The join is an inner join or a semijoin, and the cost of reading the
  first row from the probe input is lower than from the build input.

This commit implements this change. Note that for semijoin and inner
join, this change is only implemented for Hypergraph. Cost estimates
for the traditional optimizer are incomplete, and we are reluctant to
change the existing behavior.

Change-Id: I55682635ebb581f8b18ba9f08de60944b4a3c6fe