Bug#35471471 MySQL Server crashes when executing query
When pushing ORDER BY down into a redundant nested block, we fail to
increment these two counters with the contributions from the ORDER BY.

   Query_block::n_sum_items
   Query_block::select_n_where_fields

This causes the estimate the max size of base_ref_items allocated in
Query_block::setup_base_ref_items to be too low, and we see the crash
when trying to add more elements during split_sum_func (due to
presence of window function).

Solution: when pushing down ORDER BY, increment the counters.

We can check that the counters are the same with and without the extra
nesting of the repro by activating "info" trace and we should see

info: setup_ref_array this 0x...    9 :    1    0    1    1    4    2    0

for both cases with the patch.

Change-Id: Ie83653cd371b4c2732985a56ef93a53228f25e1c