Bug#33989165 MySQL Router should not disconnect client connections upon adding a member
When the user adds a node to the cluster (cluster.addInstance()) there is a
short period of time, when this node is present in the GR metadata but still not
present in the cluster metadata. This breaks Router assumptions, as the function
that calculates quorum assumes this will never happen. This assumption is wrong
and it is not satisfied in this case. The Router treat this scenario as no
quorum (even thought the new node is ONLINE) and  drops all the existing
connections. It brings the service back once the new node appears in the cluster
metadata.

This patch fixes this behavior. The GR quorum no longer depends on cluster
metadata, only on what GR is reporting. Once the quorum decision is made,
and there is a quorum, the Router uses the nodes that are present in
both GR and cluster metadata as a valid destinations.

Change-Id: I1831acc553678d453bca63b6c0a7dd8f705c9bf0