Bug#35471522 AccessPath* MoveCompositeIteratorsFromTablePath(AccessPath*, const Query_block& Bug#35228083 Assertion `false' failed|sql/sql_executor.cc
Assert when ordering a unary (nested block) and the order by contains a
window function:

   ( block LIMIT .. ) ORDER BY .. window-function ..

The ORDER BY is somewhat conservatively not pushed into the nested
block in this case due to the presence of the LIMIT.

MoveCompositeIteratorsFromTablePath is missing a case for
bottom_of_table_path->type being a window, so we see an assert in
debug build. In production build I do not see a crash.  The patch adds
handling of this case which makes the repro not assert in debug
builds.

For explain, we saw another assert. To fix this, we added a new case
calling a new function EstimateWindowCost.

The issue also contains another repro which does not contain a window
function which fails on the same assert. This is a similar bug
but a distinct error: in the case the unexpected access path is a
DYNAMIC_INDEX_RANGE_SCAN. The patch also solves this issue.

Finally, there is another issue 35228083 which is a duplicate of the
first repro (window function), so this patch solves that as well.

Change-Id: I2cea641a3206dfdc276e85583c624c3c1e7b26a2