WL#16455 Triggers should not be loaded on SELECTs
For read-only statements, avoid caching fully-loaded/parsed
trigger bodies in the table cache.

This is done by implementing lazy full-loading/parsing of trigger
bodies. Instead of always parsing bodies and finalizing their
loading, we only do so for operations that modify data. Once such
an operation completes, we return the TABLE instance its associated
triggers to the table cache. We then try to re-use TABLE instances
with fully-loaded triggers for operations that modify data, and
to re-use TABLE instances without trigger bodies for read-only
operations. To achieve that, the table cache code was changed
to keep two lists of unused TABLE objects for each table cache
element rather than one -- with and without fully-loaded triggers,
respectively.

To ensure that a spike in data change workload won't fill up the table
cache with TABLE objects with fully-loaded triggers, a soft limit on
the number of such objects in the cache is introduced. If this limit
is exceeded, the least recently used TABLE objects with
fully-loaded/parsed triggers are evicted from the cache.

Users may control this soft limit using new global variable,
table_open_cache_triggers. Its default value is 512*1024 so
that eviction of unused TABLE objects with fully-loaded
triggers happens as before this patch.

Furthermore, new status variables were added to allow users
to see how well TABLE objects with fully-loaded triggers are
cached, and whether table_open_cache_triggers needs to be
adjusted. The new status variables are:

- table_open_cache_triggers_hits
- table_open_cache_triggers_misses
- table_open_cache_triggers_overflows

Static meta-data that can be shared between instances of a trigger
is moved to TABLE_SHARE. This further saves memory. It also speeds
up trigger initialization, since we now perform the first phase of
trigger loading (the reading of trigger information from the
data-dictionary or its cache, and the filling-in of the static
attributes of trigger objects) only once. This happens during
TABLE_SHARE creation/in open_table_def(), rather than when each
TABLE object is constructed from TABLE_SHARE.

These shared objects and their string attributes are allocated
on the share's memory root. They can not be used directly for
trigger execution, but are used to create the trigger objects
on the TABLE instances which can be.

Since parse errors of triggers are rare, this patch also reduces
memory consumption by not statically allocating a buffer for
error messages in the Table_trigger_dispatcher and Trigger classes.
Instead, variable-size buffers are allocated on the memory root on
demand (i.e. when a parse error actually occurs).

Minor change in behavior introduced by this patch:

SHOW CREATE TRIGGER now shows the definition even of triggers that
have unparseable bodies. Such trigger bodies may be found in the
data-dictionary after upgrading the server to a version that no longer
supports the syntax used in those triggers.

Contributed by Dmitry Lenev.

Change-Id: I050a63e9e6d94240e3817e4cdadfab99148e375d