Bug #33398681 - ibd tablespace file is moved back from innodb_directories to datadir after ALTER TABLE
In case of ALTER TABLE .. FORCE and ALTER TABLE .. ENGINE, tables are
recreated.  The default location for recreated tables is default data
dir.  If the table was created with Data Directory clause, the table
is created in the same source directory.

Now, If the source table was moved to a different location and then
the above Alter statements are run, the table gets recreated in the
default data dir. This does not seem correct as the user would expect
it to be recreated in the original directory. Furthermore, ALTER TABLE
statement does not allow change in data directory so it is important
for the table to be recreated at the same location.

The solution is to update the data directory flag for corresponding
dd_table during restart after ibd file was just moved to new location,
get the current path of that ibd and reuse this path during
recreation.

Also, for those ibd files which were moved during previous versions of
the server, we recognize them also as moved by checking the following
conditions in fil_tablespace_path_equals():
1. The file must have .ibd extension
2. old_path and new_path must point to same location (because in
previous version when the file is just moved, old_path is updated to
new location already)
3. new_path must be different from default data directory
4. dd_table should already exist for this tablespace
5. dd_table data directory flag doesn't exist (because in prev versions
before this fix, the dd_table data directory flag is not updated for
those ibd files that are moved. This check is added because for those
tables that are created with data directory clause, this flag is true.
We want to eliminate such cases and be sure that the ibd file is moved)

For such ibd files which satisfy the above conditions, we recognize
them as moved and populate their tablespace information separately in
m_moved_previously_before_90000 which is later used to update dd_table.
Note that the variable is named with version 9.0.0 as this is needed
for data directories before 9.0.0. We can remove this code in 10.0.0.

We added a common method
dd_update_table_and_partitions_after_dir_change() which updates
dd_table data directory flag for moved ibd files in above cases.

Change-Id: I5d4fc2e1f7d23f6a8a9c152449d9d5ecf2344833