Bug#35612113  Bootstrap and metadata-cache data models in the trace files differ.
The bootstrap and the metadata-cache component tests set the metadata
in different way in tracefiles (different json field names etc.)
There is no reason for a difference there as this is the same metadata
schema both those test groups are setting.
It is hard to follow, maintain and extend.

This patch unifies the data set by the bootstrap and metadata-cache component
tests in the tracefiles for mock servers.

Change-Id: Iaedc9368cdb83c86a88e965f03c7e19a99883741