Bug#35432973 InnoDB: processing single character tokens with FTS parser plugin
In scenarios where FTS parsing plugin (like ngram) is used with single-character tokens,
incorrect behavior occurs due to the fact that fts_tokenize_ctx_t::processed_len has
double meaning: normally for built-in parser it indicates how much of the document has
been processed, but when parser plugin was used and the document was not completely
processed, a non-zero value (which in practice was 1) would indicate that tokenization
by the plugin was complete. However the surrounding code did not take into account
this change of semantics - this is amended.

Patch based on contribution by Shaohua Wang

Change-Id: I752f6cc57b9db38728146f4543b59183b7bb258c