BUG#35616866 MYSQL, pass cmd line to TELEMETRY_CLIENT plugin
Problem
=======

Starting with MySQL 8.1, the mysql client can load a telemetry_client
plugin, which has its own configuration parameters.

When providing configuration parameters for the plugin
in the plugin configuration file, the plugin can see the configuration and
works properly.

When providing configuration parameters for the plugin
in the command line argument that invoke the mysql client however,
these options are rejected as unknown (by the client),
instead of being given to the plugin.

Root cause
==========

The code that process options in the mysql client is not expecting
to have plugins, and processes all options at once.

Fix
===

Split options processing in the client into smaller phases.

First, options known to mysql proper are parsed,
while unknown options (to mysql) are preserved.

Second, when a telemetry_client plugin is to be loaded,
remaining unknown options are given to the plugin,
which consumes the options it supports.

Third, any remaining options, not consumed by the plugin,
are reported as unknown.

Change-Id: I2c8fb1a2913354b1cb93e58a8b08d21268191bea